backend/requirements.txt
aiohttp>=3.8.0,<4.0.0
aiosignal>=1.3.0,<2.0.0
annotated-types>=0.6.0,<1.0.0
anyio>=3.7.0,<4.0.0
assemblyai>=0.19.0,<1.0.0
async-timeout>=4.0.0,<5.0.0
attrs>=23.1.0,<24.0.0
certifi>=2023.7.0,<2024.0.0
charset-normalizer>=3.3.0,<4.0.0
click>=8.1.0,<9.0.0
dataclasses-json>=0.6.0,<1.0.0
dnspython>=2.4.0,<3.0.0
fastapi>=0.104.0,<0.110.0
frozenlist>=1.4.0,<2.0.0
gitdb>=4.0.0,<5.0.0
GitPython>=3.1.0,<4.0.0
h11>=0.14.0,<1.0.0
httpcore>=1.0.0,<2.0.0
httptools>=0.6.0,<1.0.0
httpx>=0.25.0,<1.0.0
idna>=3.4,<4.0
jsonpatch>=1.33,<2.0
jsonpointer>=2.4,<3.0
langchain>=0.1.0,<0.2.0
langsmith>=0.0.77,<0.1.0
loguru>=0.7.0,<1.0.0
marshmallow>=3.20.0,<4.0.0
multidict>=6.0.0,<7.0.0
mypy-extensions>=1.0.0,<2.0.0
numpy>=1.26.0,<2.0.0
openai>=1.3.0,<2.0.0
packaging>=23.2,<24.0
weaviate-client>=4.16.0,<5.0.0
pydantic>=2.5.0,<3.0.0
pydantic_core>=2.10.0,<3.0.0
python-dateutil>=2.8.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
PyYAML>=6.0.0,<7.0.0
requests>=2.31.0,<3.0.0
six>=1.16.0,<2.0.0
smmap>=5.0.0,<6.0.0
sniffio>=1.3.0,<2.0.0
SQLAlchemy>=2.0.0,<3.0.0
starlette>=0.27.0,<1.0.0
tenacity>=8.2.0,<9.0.0
tqdm>=4.66.0,<5.0.0
typing-inspect>=0.9.0,<1.0.0
typing_extensions>=4.8.0,<5.0.0
urllib3>=2.0.0,<3.0.0
uvicorn>=0.27.0,<1.0.0
uvloop>=0.19.0,<1.0.0
watchfiles>=0.21.0,<1.0.0
websockets>=12.0,<13.0
yarl>=1.9.0,<2.0.0
